clean up keys; use existing pem if present; accept args

+++ proxy/stunnel.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
-SVC=stunnel
-LISTEN=8443
-DEST=8080
+LISTEN=${1:-8443}
+DEST=${2:-8080}
+SVC=${3:-stunnel}
 
 echo > ${SVC}.conf "
 foreground = yes
@@ -13,14 +13,18 @@ connect = $DEST
 cert = ${SVC}.pem
 "
 
-touch ${SVC}.key ${SVC}.crt ${SVC}.pem
-chmod 600 ${SVC}.key ${SVC}.crt ${SVC}.pem
+if [ ! -f ${SVC}.pem ]; then
+	touch ${SVC}.key ${SVC}.crt ${SVC}.pem
+	chmod 600 ${SVC}.key ${SVC}.crt ${SVC}.pem
 
-# openssl genrsa -out isn't playing nice ... wth?
-# FIXME: use genpkey
-openssl genrsa 4096 > ${SVC}.key
-# params on cmdline?
-yes '' | openssl req -new -key ${SVC}.key -x509 -days 90 -out ${SVC}.crt
-cat ${SVC}.key ${SVC}.crt > ${SVC}.pem
+	# openssl genrsa -out isn't playing nice ... wth?
+	# FIXME: use genpkey
+	openssl genrsa 4096 > ${SVC}.key
+	# params on cmdline?
+	yes '' | openssl req -new -key ${SVC}.key -x509 -days 90 -out ${SVC}.crt
+	cat ${SVC}.key ${SVC}.crt > ${SVC}.pem
+	rm ${SVC}.key ${SVC}.crt
+	trap "rm ${SVC}.pem" EXIT
+fi
 
 stunnel ${SVC}.conf
